% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speech2text_batch_api.R
\name{create_transcription}
\alias{create_transcription}
\title{Create a transcription by sending audio files.}
\usage{
create_transcription(az_cognitive_service_speechservices_obj,
  transcription_body = NULL, audio_folder_SAS, diarizationEnabled = FALSE,
  wordLevelTimestampsEnabled = FALSE,
  punctuationMode = "DictatedAndAutomatic", profanityFilterMode = "Masked",
  locale = "de-AT", displayName = "German Transcription for Austria")
}
\arguments{
\item{az_cognitive_service_speechservices_obj}{an speech service object of class az_cognitive_service}

\item{transcription_body}{you can even create your own request body by using the function create_transcription_req_body}

\item{audio_folder_SAS}{This is used for sending data}
}
\value{
the transcription id and operation that is to be used in next step for get_transcription_status
}
\description{
Create a transcription by sending audio files.
}
\details{
This function creates a transcription event in azure and returns you the transcription ID that should be used again.
}
