% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speech2text_batch_api.R
\name{create_transcription_req_body}
\alias{create_transcription_req_body}
\title{Create a request body for speech transcription}
\usage{
create_transcription_req_body(audio_folder_SAS, diarizationEnabled = FALSE,
  wordLevelTimestampsEnabled = FALSE,
  punctuationMode = "DictatedAndAutomatic", profanityFilterMode = "Masked",
  locale = "en-US",
  displayName = "Transcription from azure directly within R programming",
  ...)
}
\arguments{
\item{audio_folder_SAS}{A URL with Shared access Signature for azure storage Where audio files are stored.}

\item{diarizationEnabled}{Boolean parameter for speech recognition.}

\item{punctuationMode}{defining punctuation options see speechtotext batch api for more options}

\item{profanityFilterMode}{defining profanityFilterMode options see speechtotext batch api for more options}

\item{locale}{defining language options see speechtotext batch api for more options}

\item{displayName}{displayName is a text that appears with json more like a comment}

\item{...}{pass additional parameter either as a named vector or a named list}
}
\value{
It returns a body that is needed to be sent in the request.
}
\description{
Create a request body for speech transcription
}
\details{
This function creates a body for speech to text api for creating transcription.
}
