% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speech2text_batch_api.R
\name{extract_file_name}
\alias{extract_file_name}
\title{Extract name filename from file url string}
\usage{
extract_file_name(file_data, start_word, end_word)
}
\arguments{
\item{transcription_files_response}{a A string contains the whole url of file that was sent for transcription}
}
\value{
It returns the filesnames from string
}
\description{
Extract name filename from file url string
}
\details{
Extract file names from file_data obtained from url
}
