% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cognitive_token.R
\name{get_cognitive_token}
\alias{get_cognitive_token}
\title{Obtain authentication token for a cognitive service}
\usage{
get_cognitive_token(key, region = "global", token_url = NULL)
}
\arguments{
\item{key}{The subscription key for the service.}

\item{region}{The Azure region where the service is located.}

\item{token_url}{Optionally, the URL for the token endpoint.}
}
\description{
Obtain authentication token for a cognitive service
}
