% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speech2text_batch_api.R
\name{get_trans_id}
\alias{get_trans_id}
\title{Extracts transcription ID from payload sent by create transcription API}
\usage{
get_trans_id(create_transcription_data)
}
\arguments{
\item{create_transcription_data}{response you recieved from create transcription API}
}
\value{
It returns a transcription id and the url for next operation to be used to get status
}
\description{
Extracts transcription ID from payload sent by create transcription API
}
\details{
This function returns the transcription id and operation that has to be used in subsequent function
}
