% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speech2text_batch_api.R
\name{get_transcription_files}
\alias{get_transcription_files}
\title{Create a transcription by sending audio files.}
\usage{
get_transcription_files(az_cognitive_service_speechservices_obj,
  create_transcription_data)
}
\arguments{
\item{az_cognitive_service_speechservices_obj}{an speech service object of class az_cognitive_service}

\item{create_trans_data}{the data you get from create_transcription function}
}
\value{
It returns a body that is needed to be sent in the request.
}
\description{
Create a transcription by sending audio files.
}
\details{
This function creates a transcription event in azure and returns you the transcription ID that should be used again.
}
