% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speech2text_batch_api.R
\name{get_transcription_files_response}
\alias{get_transcription_files_response}
\title{Create a transcription by sending audio files.}
\usage{
get_transcription_files_response(az_cognitive_service_speechservices_obj,
  create_transcription_data)
}
\arguments{
\item{az_cognitive_service_speechservices_obj}{an speech service object of class az_cognitive_service}

\item{create_trans_data}{the data you get from create_transcription function}
}
\value{
It returns the response you get from Transcription files API
}
\description{
Create a transcription by sending audio files.
}
\details{
This function asks for files after transcription event is done running job
}
