% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speech2text_batch_api.R
\name{get_transcription_files_url}
\alias{get_transcription_files_url}
\title{Create a transcription by sending audio files.}
\usage{
get_transcription_files_url(transcription_files_response)
}
\arguments{
\item{transcription_files_response}{a file url response you get from get_transcription_files_response function}
}
\value{
It returns the transcribed files data
}
\description{
Create a transcription by sending audio files.
}
\details{
This function gives you the output of all the files
}
